visualization/requirements.txt
keymap-drawer==1.0.2
PyYAML==6.0